[APPROVE:1006] v1.6.6网络诊断版：添加NetworkDiagnostics，详细排查网络问题

+++ app/src/main/java/com/honghuang/guard/VoiceWakeManager.java
@@ -148,7 +148,7 @@ public class VoiceWakeManager {
         // 或使用火山引擎提供的特征提取API
         
         // 临时返回简化特征
-        return new float[]{calculateEnergy(audioData)};
+        return new float[]{(float) calculateEnergy(audioData)};
     }
     
     /**
